revising Excel tasks
