+++ DataTransformation/DataTransformation.csproj
@@ -119,8 +119,9 @@
     <Compile Include="Mongo\IMongoDB.cs" />
     <Compile Include="Office\ExcelFormats.cs" />
     <Compile Include="Office\ExcelColumn.cs" />
-    <Compile Include="Office\ExcelDestination.cs" />
+    <Compile Include="Office\ExcelWorkbookDestination.cs" />
     <Compile Include="Mail\SendEmailTask.cs" />
+    <Compile Include="Office\ExcelSheetTransformation.cs" />
     <Compile Include="SqlServer\MultiExecuteSqlTask.cs" />
     <Compile Include="SqlServer\RecordReader.cs" />
     <Compile Include="SqlServer\TableMetaDataCache.cs" />
